+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneMobileAppsMicrosoftStoreAppWindows10/MSFT_IntuneMobileAppsMicrosoftStoreAppWindows10.psm1
@@ -229,6 +229,7 @@ function Get-TargetResource
         $assignmentResult = @()
         if ($assignmentsValues.Count -gt 0)
         {
+            [array]$assignmentsValues = $assignmentsValues | Where-Object -FilterScript { $_.source -eq 'direct' }
             $assignmentResult += ConvertFrom-IntuneMobileAppAssignment -Assignments $assignmentsValues -IncludeDeviceFilter $true
         }
         $results.Add('Assignments', $assignmentResult)