+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* AADTenantAppManagementPolicy
+  * Fixed an issue where the `RestrictForAppsCreatedAfterDateTime` timestamp
+    was incorrectly converted to a DateTime object.
+    FIXES [#7291](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7291)
+* IntuneDeviceCompliancePolicyMacOS
+  * Fixed an issue where `ScheduledActionsForRule` was not compared correctly.
+* MISC
+  * Fixed an issue where assignments through a policy set were exported for
+    resources that were not of the `IntunePolicySets` type.
+
 # 1.26.701.1
 
 * AADCrossTenantAccessPolicyConfigurationPartner